Finish test algorithms page

+++ MetaheuristicsAPI/MetaheuristicsAPI/Schemas/FitnessFunctionSchema.cs
@@ -1,4 +1,4 @@
 ﻿namespace MetaheuristicsAPI.Schemas
 {
-    public record FitnessFunctionSchema(string FunctionName, bool AnyDims, double[]? Dims = null);
+    public record FitnessFunctionSchema(string FunctionName, bool AnyDims, int[]? Dims = null);
 }
