+++ MetaheuristicsPlatform/MetaheuristicsPlatform/schemas/FitnessFunctionSchema.cs
@@ -1,4 +1,4 @@
 ﻿namespace MetaheuristicsPlatform.Schemas
 {
-    public record FitnessFunctionSchema(string FunctionName, bool AnyDims, double[]? Dims = null);
+    public record FitnessFunctionSchema(string FunctionName, bool AnyDims, int[]? Dims = null);
 }
